create common TI interface in prinz
